fixed integer prices and coorect orders list depending on the user role

+++ src/main/java/com/elearn/fp/db/entity/Item.java
@@ -80,7 +80,7 @@ public class Item {
         return productPrice;
     }
 
-    public void setProductPrice(int productPrice) {
+    public void setProductPrice(Double productPrice) {
         this.productPrice = productPrice;
     }
